# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gerenciamento de bancos de dados.",
        "title": "Database Manager API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8181",
    "basePath": "/api/",
    "paths": {
        "/connection": {
            "get": {
                "description": "Retorna todas as conexões cadastradas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "connection"
                ],
                "summary": "Lista conexões salvas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ConnectionConfig"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "connection"
                ],
                "summary": "Altera conexão",
                "parameters": [
                    {
                        "description": "Dados da conexão",
                        "name": "connection",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ConnectionConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "connection"
                ],
                "summary": "Salva conexão",
                "parameters": [
                    {
                        "description": "Dados da conexão",
                        "name": "connection",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ConnectionConfig"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    }
                }
            }
        },
        "/connection/{id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "connection"
                ],
                "summary": "Remove conexão",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da conexão",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    }
                }
            }
        },
...
        "/table/column": {
            "put": {
                "description": "Altera uma coluna em uma tabela",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "column"
                ],
                "summary": "Altera coluna",
                "parameters": [
                    {
                        "description": "Coluna",
                        "name": "column",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ColumnRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Cria uma coluna em uma tabela",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "column"
                ],
                "summary": "Cria coluna",
                "parameters": [
                    {
                        "description": "Coluna novo",
                        "name": "column",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ColumnRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    }
                }
            }
        },
        "/table/column/{table}": {
            "get": {
                "description": "Retorna uma lista com as colunas e o schema de uma tabela",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "column"
                ],
                "summary": "Listar colunas de uma tabela",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tabela",
                        "name": "table",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Table"
                        }
                    }
                }
            }
        },
        "/table/column/{table}/{column}": {
            "delete": {
                "description": "Remove uma coluna de uma tabela",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "column"
                ],
                "summary": "Remove coluna",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tabela",
                        "name": "table",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Coluna",
                        "name": "column",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Column": {
            "type": "object",
            "properties": {
                "autoIncrement": {
                    "type": "boolean",
                    "example": false
                },
                "defaultValue": {
                    "type": "string",
                    "example": ""
                },
                "length": {
                    "type": "integer",
                    "example": 10
                },
                "name": {
                    "type": "string",
                    "example": "teste"
                },
                "nullable": {
                    "type": "boolean",
                    "example": true
                },
                "primary": {
                    "type": "boolean",
                    "example": false
                },
                "type": {
                    "type": "string",
                    "example": "INT"
                },
                "unique": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.ColumnRequest": {
            "type": "object",
            "properties": {
                "column": {
                    "$ref": "#/definitions/models.Column"
                },
                "oldName": {
                    "type": "string"
                },
                "table": {
                    "type": "string"
                }
            }
        },
        "models.ConnectionConfig": {
            "type": "object",
            "properties": {
                "database": {
                    "type": "string",
                    "example": ""
                },
                "driver": {
                    "type": "string",
                    "example": "mysql"
                },
                "host": {
                    "type": "string",
                    "example": "127.0.0.1"
                },
                "id": {
                    "type": "string",
                    "example": ""
                },
                "name": {
                    "type": "string",
                    "example": ""
                },
                "password": {
                    "type": "string",
                    "example": "senha123"
                },
                "port": {
                    "type": "integer",
                    "example": 3306
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
...
        "models.Table": {
            "type": "object",
            "properties": {
                "columns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Column"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "teste"
                }
            }
        },
...
        "respond.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}